requirements.txt
# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.25.0

# Database & Storage
neo4j>=5.0.0
python-dotenv>=1.0.0
aiofiles>=24.1.0

# Communication & APIs
twilio>=9.0.0
python-multipart>=0.0.20
httpx>=0.25.0

# Performance & Monitoring  
psutil>=5.9.0

# Optional Dependencies
# Uncomment if needed for advanced features:
# spacy>=3.7.0              # For advanced NLP
# beautifulsoup4>=4.12.0     # For web scraping  
# playwright>=1.40.0         # For browser automation
# openai>=1.0.0              # For OpenAI integration
# jinja2>=3.1.0              # For templating